Added feature of keeping current counts of each type of entity using static variables and functions.
Added destructors to each entity type and added calls to delete those entities when they die (in Map::updateEntities().

Added feature to updateEntities() that checks current cow and grass populations.  If it's under some number then new entities will automatically be generated in a random location on the map in order to prevent extinction.